Use reusabel MockLogger to capture warning messages
